Population, Prices, and Quantities for 60 countries (1980).

A dataset with prices and quantities for 11 goods in 60 countries 1980. Prices and expenditure were collected around the world by the International Comparison Project, and this is the rescaled dataset such that all US prices are set to unity, and quantities are measured as expenditues in USD.

Neary2004

Format

A data frame with 60 rows and 24 variables:

country

Name of country (chr)

pop1980

1980 population in country (in thousands)

p1

Price of food

p2

Price of beverages

p3

Price of tobacco

p4

Price of clothing and footwear

p5

Price of gross rents

p6

Price of fuel and power

p7

Price of house furnishings

p8

Price of medical care

p9

Price of transport and communication

p10

Price of recreation and education

p11

Price of miscellaneous

q1

Quantity of food (per capita)

q2

Quantity of beverages (per capita)

q3

Quantity of tobacco (per capita)

q4

Quantity of clothing and footwear (per capita)

q5

Quantity of gross rents (per capita)

q6

Quantity of house furnishings (per capita)

q7

Quantity of house furnishings (per capita)

q8

Quantity of medical care (per capita)

q9

Quantity of transport and communications (per capita)

q10

Quantity of recreating and education (per capita)

q11

Quantity of miscellaneous (per capita)

Source

From the paper Neary, J Peter. 2004. "Rationalizing the Penn World Table: True Multilateral Indices for International Comparisons of Real Income." American Economic Review, 94 (5): 1411-1428. https://www.aeaweb.org/articles?id=10.1257/0002828043052286 Data collected at http://users.ox.ac.uk/~econ0211/gaia/gaia.htm